Fix some positioning and sizing
13pt was too big on smaller screens (1366x768) so I added a special case where if your screen is that small or smaller it the font size (and subsequently the size of everything else) will be 12pt instead.

The left card was smaller than the right card after I added padding to the right and not the left. That is now fixed, but I had to add some negative margins and calc() statements to left's image sizing to make it still work.

The name banner was wonky. I changed some of its magic numbers and it looks how it should again.